refactor: deck-list-page handlers refactor
