+++ src/components/ui/layout-components/tables/deck-list-table/deck-list-table.tsx
@@ -7,24 +7,25 @@ import { Actions } from '@/components/ui/layout-components/actions'
 import { TableBody, TableContainer, TableHeader, TableRow } from '@/components/ui/primitives'
 import { Deck } from '@/services/decks/deck.types'
 import { DIALOG_ACTION, DIALOG_ENTITY, PATH, VARIANT } from '@/shared/enums'
+import { useSearchParamUpdater } from '@/shared/hooks'
 import { convertToDDMMYYYY } from '@/shared/utils/convert-date-ddmmyyyy'
 
 import { HeaderCell, PositionCell } from '../container-components'
 
 type DecksListTableProps = {
   decks: Deck[]
-  onSort: (orderBy: string) => void
 }
 
-export const DeckListTable = ({ decks, onSort }: DecksListTableProps) => {
+export const DeckListTable = ({ decks }: DecksListTableProps) => {
   const [deckId, setDeckId] = useState('')
   const [showEditDeckDialog, setShowEditDeckDialog] = useState(false)
   const [showDeleteDeckDialog, setShowDeleteDeckDialog] = useState(false)
   const [sortId, setSortId] = useState('')
+  const updateSearchParam = useSearchParamUpdater()
 
   const sortHandler = (orderBy: string, sortId: string) => {
     setSortId(sortId)
-    onSort(orderBy)
+    updateSearchParam({ currentPage: 1, orderBy })
   }
 
   const deckData = decks.find(el => el.id === deckId) ?? ({} as Deck)
